Introduce 'Messager' component
